+++ frontend/src/features/tasks/components/EditTaskModal.tsx
@@ -1,8 +1,8 @@
-import "./EditTaskModal.css"
+import "../../../components/EditTaskModal.css"
 import { useEffect, useState } from "react"
-import { updateTask } from "../lib/api"
-import type { Task } from "../types/task"
-import { useApiError } from "../hooks/useApiError";
+import { updateTask } from "../../../lib/api"
+import type { Task } from "../../../types/task"
+import { useApiError } from "../../../hooks/useApiError";
 import { useNavigate } from "react-router-dom";
 
 type Props = {
