+++ frontend/src/features/tasks/components/TaskList.tsx
@@ -4,7 +4,7 @@ import { useEffect, useState, useCallback } from "react";
 import type { Task } from "../../../types/task";
 import { getMe, fetchTasks, toggleTaskComplete, deleteTask } from "../../../lib/api";
 import { TaskAdd } from "./TaskAdd";
-import EditTaskModal from "../../../components/EditTaskModal"
+import EditTaskModal from "./EditTaskModal"
 import { Navigate, useNavigate } from "react-router-dom"
 import { useApiError } from "../../../hooks/useApiError";
